+++ mock-infrastructure/src/main/java/com/sequenceiq/mock/clouderamanager/base/batchapi/AbstractBatchApiHandler.java
@@ -1,7 +1,5 @@
 package com.sequenceiq.mock.clouderamanager.base.batchapi;
 
-import java.math.BigDecimal;
-
 import org.springframework.http.HttpStatus;
 
 import com.sequenceiq.mock.swagger.model.ApiBatchResponseElement;
@@ -10,7 +8,7 @@ public abstract class AbstractBatchApiHandler implements BatchApiHandler {
 
     protected ApiBatchResponseElement getSuccessApiBatchResponseElement(String response) {
         return new ApiBatchResponseElement()
-                .statusCode(BigDecimal.valueOf(HttpStatus.OK.value()))
+                .statusCode(HttpStatus.OK.value())
                 .response(response);
     }
